refactor: reorganize project structure
- Move all source code to src/ directory
- Reorganize config, docker, scripts, and docs into separate directories
- Update import paths and Docker configurations
- Improve deployment script paths and environment handling
