+++ config/nodemon.json
