+++ docs/README.md
@@ -1,0 +1,24 @@
+# GitTrack Bot - Development Documentation
+
+This directory contains development and contribution documentation for the GitTrack Discord Bot.
+
+## Documentation Files
+
+- **[README.md](../README.md)** - Main project documentation and setup guide
+- **[CONTRIBUTING.md](CONTRIBUTING.md)** - How to contribute to the project
+- **[BRANCH_PATTERNS.md](BRANCH_PATTERNS.md)** - Understanding branch filtering
+
+
+## Project Organization
+
+- **`src/`** - All source code
+  - `commands/` - Discord bot slash commands
+  - `functions/` - Utility functions and helpers
+  - `handlers/` - Webhook and event handlers
+  - `bot.js` - Main bot logic
+  - `index.js` - Application entry point
+- **`prisma/`** - Database schema and migrations
+- **`config/`** - Configuration files (nodemon.json, etc.)
+- **`docker/`** - Docker deployment files
+- **`scripts/`** - Deployment and utility scripts
+- **`docs/`** - Documentation (this directory)
