+++ scripts/deploy.sh
@@ -5,7 +5,15 @@
 
 set -e
 
+# Get the directory where this script is located
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+# Get the project root directory (parent of scripts)
+PROJECT_ROOT="$(dirname "$SCRIPT_DIR")"
+
 echo "🚀 Deploying GitTrack Discord Bot..."
+
+# Change to project root directory
+cd "$PROJECT_ROOT"
 
 # Check if .env file exists
 if [ ! -f .env ]; then
@@ -31,9 +39,9 @@ echo "✅ Environment variables validated"
 
 # Build and start the services
 echo "🔨 Building and starting services..."
-docker-compose -f docker-compose.dev.yml down
-docker-compose -f docker-compose.dev.yml build --no-cache
-docker-compose -f docker-compose.dev.yml up -d
+docker-compose -f docker/docker-compose.dev.yml down
+docker-compose -f docker/docker-compose.dev.yml build --no-cache
+docker-compose -f docker/docker-compose.dev.yml up -d
 
 # Wait for services to be ready
 echo "⏳ Waiting for services to be ready..."
@@ -41,7 +49,7 @@ sleep 10
 
 # Setup database schema
 echo "🗄️ Setting up database schema..."
-docker-compose -f docker-compose.dev.yml exec bot sh -c "
+docker-compose -f docker/docker-compose.dev.yml exec bot sh -c "
   if [ -d 'prisma/migrations' ] && [ \"\$(ls -A prisma/migrations)\" ]; then
     echo 'Using migrations...'
     npx prisma migrate deploy
@@ -53,8 +61,8 @@ docker-compose -f docker-compose.dev.yml exec bot sh -c "
 
 # Check if services are running
 echo "🔍 Checking service status..."
-docker-compose -f docker-compose.dev.yml ps
+docker-compose -f docker/docker-compose.dev.yml ps
 
 echo "✅ GitTrack Discord Bot deployed successfully!"
 echo "🌐 Bot is running on http://localhost:3000"
-echo "📊 Check logs with: docker-compose -f docker-compose.dev.yml logs -f bot" 
+echo "📊 Check logs with: docker-compose -f docker/docker-compose.dev.yml logs -f bot"
