+++ src/commands/ping.js
