+++ src/commands/remove-repo.js
