+++ src/commands/setup.js
