+++ src/commands/status.js
