+++ src/functions/limitChecker.js
