+++ src/functions/permissionChecker.js
