+++ src/handlers/webhookHandler.js
@@ -3,8 +3,8 @@ const crypto = require('crypto');
 const querystring = require('querystring');
 const { handleMilestoneEvent, handleWorkflowRunEvent } = require('./milestoneAndWorkflowHandlers');
 const { handlePRReviewEvent, handlePRReviewCommentEvent } = require('./pullRequestHandlers');
-const { checkChannelLimit } = require('./functions/limitChecker');
-const { findMatchingBranches } = require('./functions/branchMatcher');
+const { checkChannelLimit } = require('../functions/limitChecker');
+const { findMatchingBranches } = require('../functions/branchMatcher');
 
 function initializeWebServer(prisma, botClient) {
   const app = express();
